move label to component
