+++ platform/src/lib.rs
@@ -60,6 +60,7 @@ pub mod components {
     pub mod checkbox;
     pub mod icons;
     pub mod input;
+    pub mod label;
     pub mod pi_graph;
     pub mod popup;
     pub mod select;
